server: split off create_* from ldap_handler
