+++ server/src/infra/ldap/mod.rs
@@ -1,4 +1,5 @@
 pub mod compare;
+pub mod create;
 pub mod handler;
 pub mod password;
 pub mod search;
